requirements.txt
django==3.1.1
pandas==0.25.3
requests==2.18.4
django-haystack==3.0

gunicorn==19.9.0
prometheus-client==0.7.1
